Bump nixpkgs-unstable, ahmed: use unstable PaperMC
I needed PaperMC v1.24.1 which happens to be what is packaged in
NixOS/nixpkgs@852ff1d9e.

I really ought to figure out a way to decouple the version of Minecraft
from the nixpkgs version. I don't want to be in the situation where I
want to bump unstable, but don't want to switch to a new Minecraft
version.